Update test files by moving them into test folder
