+++ src/tests/calculate.test.js
@@ -1,4 +1,4 @@
-import calculate from './calculate';
+import calculate from '../logic/calculate';
 
 describe('Test button fonctionalities', () => {
   test('AC button clears the calculator data object', () => {
